requirements.txt
Django>=5.2,<6.0
djangorestframework>=3.15
django-cors-headers>=4.0
requests>=2.32
